Forgot to include new Wall class file.
Also, updated path for opening the wall photo in default image viewer to the correct new path.

Also, removed the minSDK as it appears we need to do some verification testing to see what versions we can actually support.

Note: Gallery View is throwing up on the emulator. Will need to investigate to see why.